LOD that works
The goal is to connect and interlink data around historical figures from
various datasets from Open Data Trento. The results is not only a
showcase of the practicality of publishing data as Linked Data, but also
an infotainment experience to raise awareness of historical figures with
local heritage while promoting tourism and investment in the Province of
Trento. The app shows the cities and places where historical figures
have lived and connecting them with local points of interest and
artistic heritage in that area.